Create PlayerControl_Caster.ino
The caster player control only has an RF transmitter (no IR transmitter). The output signals of the RF transmitter is controlled by a potentiometer and an accelerometer attached to one hand of the player. The potentiometer is connected such that it's rotation is controlled by the wrist's rotation. The acceleration of the hand while casting is proportional to the acceleration of the hand, and the position of the potentiometer determines which spell is cast.