Update PlayerControl.ino
We are currently in the debugging stage of the code, so there are a lot of print statements that will be taken out of the code. The IR and RF transmitting and receiving work; the 16-bit signals are correctly transmitted, received, and parsed. The Virtual Wire library interfered with the IRremote library's transmission (the send functions). The fix for this was to disable VirtualWire's timer 1 while the IR code was being sent.
If the functptr carriers[] is called with an argument greater than the number of functptrs it hold, the program will restart. A check for carriers to be less than carriers' size is placed in the Get_Damage() function so that there is no danger of the program resetting.