Added code to accommodate HD44780 LCD
The code has been added to accommodate an HD44790 LCD Screen. This screen currently only displays the health of the player. The updated schematic for this can be found on the site (lasersandlogic.weebly.com). The following text will be printed on the screen at startup:

   Health:
   100

The health will be updated every time a player is damaged or healed.

The liquid crystal arduino library was used to write data to the lcd. The library can be found at http://arduino.cc/en/Reference/LiquidCrystal . Make sure you have the library on your arduino ide!

The code for the color LEDs have been temporarily removed to make pin space for the led, however they will be added again to the schematic soon.